+++ packages/babel-types/test/definitions.js
@@ -1,0 +1,8 @@
+var t = require("../lib");
+
+Object.keys(t.VISITOR_KEYS).forEach(function (type) {
+  try {
+    require("./definitions/" + type.toLowerCase());
+  }
+  catch (e) {}
+});