+++ packages/babel-types/test/fixtures/identifier.js
@@ -1,0 +1,16 @@
+// spec.Identifier (e.g. BindingIdentifier).
+// value => valid (ES2015 strict).
+module.exports = [
+  ["x", true],
+
+  // IdentifierName, ReservedWord :: Keyword.
+  ["default", false],
+
+  // ReservedWord's
+  ["null", false],
+  ["true", false],
+  ["false", false],
+
+  // Not a lexical IdentifierName.
+  ["bound x", false]
+];